[Silver II] Title: A → B, Time: 96 ms, Memory: 16920 KB -BaekjoonHub

+++ 백준/Silver/16953. A → B/A → B.java
@@ -12,32 +12,32 @@ public class Main {
 		
 		int answer = -1;
 
-		List<Long> q = new LinkedList<>();
-		q.add((long) a);
-		q.add(0l);
+		List<Integer> q = new LinkedList<>();
+		q.add(a);
+		q.add(0);
 		while (!q.isEmpty()) {
 			long num = q.remove(0);
-			long count = q.remove(0);
+			int count = q.remove(0);
 			
 			long case1 = num * 2;
 			long case2 = num * 10 + 1;
 			
 			if (case1 == b || case2 == b) {
-				answer = (int)count + 1 + 1;
+				answer = count + 1 + 1;
 				break;
 			}
 			
 			if (case1 < b) {
-				q.add(case1);
+				q.add((int)case1);
 				q.add(count + 1);
 			}
 			
 			if (case2 < b) {
-				q.add(case2);
+				q.add((int)case2);
 				q.add(count + 1);
 			} 
 		}
-		
+
 		System.out.println(answer);
 	}
